requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
pydantic>=2.0.0
asyncio
aiohttp>=3.8.0

# Image analysis and search
google-search-results>=2.4.0

# LangChain ecosystem  
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
langchain-mcp-adapters>=0.1.7
langgraph>=0.1.0

# Instagram integration
instagrapi>=2.0.0
mcp>=1.9.1 